backend/config/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
psycopg2-binary==2.9.11
django-cors-headers==4.3.1
